prepare beta

+++ src/renderer/logic/file.js
@@ -16,6 +16,8 @@ import downloader from './downloader'
 import nameTest from './utils/nameTest'
 import crypto from './crypt'
 
+const remote = require('@electron/remote')
+
 const sincronizeAction = {}
 sincronizeAction[state.state.UPLOAD] = uploadState
 sincronizeAction[state.state.DOWNLOAD] = downloadState
@@ -170,6 +172,9 @@ async function sincronizeCloudFile() {
     cloudIndex[elem.key] = i++
   })
   for (const f in selectIndex) {
+    if (ConfigStore.get('stopSync')) {
+      throw Error('stop sync')
+    }
     var file = select[selectIndex[f]]
     if (file.state === state.state.IGNORE) {
       continue
@@ -224,7 +229,7 @@ async function sincronizeCloudFile() {
     f.select = true
     select.push(f)
   }
-  console.log('despues sinc cloud: ', select)
+  // console.log('despues sinc cloud: ', select)
   ConfigStore.set('updatingDB', true)
   await Database.ClearFilesSelect()
   await Database.dbInsert(Database.dbFiles, select)
@@ -269,8 +274,8 @@ async function sincronizeLocalFile() {
       const selectFile = select[FileSelect]
       const selectTime = new Date(selectFile.value.createdAt).setMilliseconds(0)
       const localTime = localFile.mtime.setMilliseconds(0)
-      console.log('selectTime: ', selectTime)
-      console.log('localTime: ', localTime)
+      // console.log('selectTime: ', selectTime)
+      // console.log('localTime: ', localTime)
 
       if (selectTime > localTime) {
         // select recent, download
@@ -304,7 +309,7 @@ async function sincronizeLocalFile() {
     )
     select[indexDict[item]].needSync = true
   }
-  console.log('despues sinc local: ', select)
+  // console.log('despues sinc local: ', select)
   ConfigStore.set('updatingDB', true)
   await Database.ClearFilesSelect()
   await Database.dbInsert(Database.dbFiles, select)
@@ -321,10 +326,14 @@ async function sincronizeFile() {
   const rootPath = await Database.Get('xPath')
   const user = (await Database.Get('xUser'))
   for (const file of select) {
-    console.log('sincroniza: ', file)
-    console.log('state: ', file.state)
+    // console.log('sincroniza: ', file)
+    // console.log('state: ', file.state)
+    if (ConfigStore.get('stopSync')) {
+      throw Error('stop sync')
+    }
+    remote.app.emit('set-tooltip', `Checking file ${file.key}`)
     if (!file.select) {
-      console.log('not select')
+      // console.log('not select')
       try {
         fs.unlink(file.key)
         await Database.dbRemoveOne(Database.dbFiles, { key: file.key })
@@ -346,7 +355,7 @@ async function sincronizeFile() {
     }
     if (!file.nameChecked) {
       if (!parentFolder || parentFolder.state === state.state.IGNORE || nameTest.invalidFileName(path.basename(file.key), rootPath)) {
-        console.log('invalid name')
+        // console.log('invalid name')
         file.state = state.state.IGNORE
         file.needSync = false
       }
@@ -354,7 +363,7 @@ async function sincronizeFile() {
       await Database.dbUpdate(Database.dbFiles, { key: file.key }, { $set: file })
     }
     if (!file.needSync || file.state === state.state.IGNORE) {
-      console.log('IGNORE')
+      // console.log('IGNORE')
       continue
     }
     try {
@@ -377,6 +386,7 @@ async function sincronizeFile() {
  */
 async function uploadFile(file, localFile, cloudFile, encryptedName, rootPath, user, parentFolder) {
   try {
+    remote.app.emit('set-tooltip', `Uploading file ${file.key}`)
     const newFileInfo = await Uploader.uploadFile(file.key, localFile, cloudFile, encryptedName, rootPath, user, parentFolder)
     if (newFileInfo && newFileInfo.fileId) {
       file.value = newFileInfo
@@ -397,6 +407,7 @@ async function uploadFile(file, localFile, cloudFile, encryptedName, rootPath, u
 */
 async function downloadFile(file, cloudFile, localFile) {
   try {
+    remote.app.emit('set-tooltip', `Downloading file to ${file.key}`)
     analytics
       .track({
         userId: undefined,
@@ -450,10 +461,10 @@ async function downloadFile(file, cloudFile, localFile) {
 }
 
 async function uploadState(file, rootPath, user, parentFolder) {
-  console.log('upload')
-  console.log(parentFolder)
+  // console.log('upload')
+  // console.log(parentFolder)
   if (!parentFolder.value || !parentFolder.value.id) {
-    console.log('no parentId')
+    // console.log('no parentId')
     return
   }
   const localFile = fs.existsSync(file.key) ? fs.lstatSync(file.key) : undefined
@@ -462,24 +473,24 @@ async function uploadState(file, rootPath, user, parentFolder) {
   let cloudFile = await Database.dbFindOne(Database.dbFilesCloud, { key: file.key })
   cloudFile = cloudFile ? cloudFile.value : null
   if (cloudFile && localFile) {
-    console.log('cloud & local exist')
+    // console.log('cloud & local exist')
     const cloudTime = new Date(cloudFile.createdAt).setMilliseconds(0)
     const localTime = localFile.mtime.setMilliseconds(0)
     if (cloudTime < localTime) {
-      console.log('local recent')
+      // console.log('local recent')
       await uploadFile(file, localFile, cloudFile, encryptedName, rootPath, user, parentFolder)
       await Database.dbUpdate(Database.dbFiles, { key: file.key }, { $set: file })
       return
     }
     if (cloudTime > localTime) {
-      console.log('cloud recent')
+      // console.log('cloud recent')
       file.state = state.state.DOWNLOAD
       await downloadFile(file, cloudFile, localFile)
       await Database.dbUpdate(Database.dbFiles, { key: file.key }, { $set: file })
       return
     }
     if (cloudTime === localTime) {
-      console.log('same')
+      // console.log('same')
       file.state = state.state.SYNCED
       file.needSync = false
       file.value = cloudFile
@@ -487,24 +498,24 @@ async function uploadState(file, rootPath, user, parentFolder) {
     }
   }
   if (!cloudFile && localFile) {
-    console.log('local exist')
+    // console.log('local exist')
     await uploadFile(file, localFile, cloudFile, encryptedName, rootPath, user, parentFolder)
     await Database.dbUpdate(Database.dbFiles, { key: file.key }, { $set: file })
     return
   }
   if (cloudFile && !localFile) {
-    console.log('cloud exist')
+    // console.log('cloud exist')
     const cloudTime = new Date(cloudFile.createdAt).setMilliseconds(0)
     const selectTime = new Date(file.value.createdAt).setMilliseconds(0)
     if (cloudTime > selectTime) {
-      console.log('cloud recent')
+      // console.log('cloud recent')
       file.value = cloudFile
       file.state = state.state.DOWNLOAD
       await downloadFile(file, cloudFile, localFile)
       await Database.dbUpdate(Database.dbFiles, { key: file.key }, { $set: file })
       return
     } else {
-      console.log('select recent')
+      // console.log('select recent')
       file.state = state.state.DELETE_CLOUD
       try {
         await removeFile(cloudFile.bucket, cloudFile.fileId)
@@ -518,7 +529,7 @@ async function uploadState(file, rootPath, user, parentFolder) {
     }
   }
   if (!localFile && !cloudFile) {
-    console.log('cloud & local not exist')
+    // console.log('cloud & local not exist')
     Database.dbRemoveOne(Database.dbFiles, { key: file.key })
   }
 }
@@ -535,8 +546,8 @@ async function downloadState(file, rootPath, user, parentFolder) {
   if (cloudFile && localFile) {
     const cloudTime = new Date(cloudFile.createdAt).setMilliseconds(0)
     const localTime = localFile.mtime.setMilliseconds(0)
-    console.log('download state cloudTime: ', cloudTime)
-    console.log('download state localTime: ', localTime)
+    // console.log('download state cloudTime: ', cloudTime)
+    // console.log('download state localTime: ', localTime)
     if (cloudTime < localTime) {
       file.state = state.state.UPLOAD
       file.value.createdAt = localTime
